restructure app for modularity.
