+++ server/controllers/userController.js
@@ -1,5 +1,6 @@
 // THIS FILE STORES ALL FUNCTIONS RELATED TO USER BEHAVIOR
-var User = require('./userModel.js'),
+var db = require('../database/dbSchema.js'),
+    Users = db.Users;
     Q    = require('q'),
     jwt  = require('jwt-simple');
