+++ server/database/dbCreds.js
@@ -3,5 +3,5 @@ module.exports= {
   password : process.env.DB_PASSWORD || 'postgres',
   database : process.env.DB_NAME || 'germTracker',
   host     : process.env.DB_HOST || 'localhost',
-  client: 'pg',
+  dialect: 'postgres',
 };
