--- server/locations/locationModel.js
@@ -1,8 +1,0 @@
-// // var bookshelf = require('../config/dbConfig');
-
-// var Location = bookshelf.Model.extend({
-//   tableName: 'location',
-//   hasTimestamps: ['createdAt', 'updatedAt']
-// });
-
-// module.exports = ('Location', Location);
