+++ server/server.js
@@ -1,10 +1,10 @@
 var express = require('express');
-var db = require('./config/dbSchema.js');
+var db = require('./database/dbSchema.js');
 
 var app = express();
 
 // configure our server with all the middleware and routing
-require('./config/middleware.js')(app, express);
+require('./config/serverConfig.js')(app, express);
 
 var port = process.env.PORT || 4568;
